requirements.txt
python-dotenv==1.0.1
streamlit==1.35.0
requests==2.31.0
google-cloud-aiplatform==1.60.0
google-auth==2.29.0